requirements.txt
python-dotenv==0.21.0
attrs==22.1.0
requests==2.28.1
requests-oauthlib==1.3.1
matplotlib==3.6.1
numpy==1.23.4
pydub==0.25.1
pyAudio==0.2.12
youtube-dl==2021.12.17
websockets==10.4
openai==0.24.0